+++ src/main/java/nextstep/subway/common/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.common.domain;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
