+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,4 +1,9 @@
-package nextstep.subway.domain;
+package nextstep.subway.line.domain;
+
+import nextstep.subway.common.domain.BaseEntity;
+import nextstep.subway.section.domain.Section;
+import nextstep.subway.section.domain.Sections;
+import nextstep.subway.station.domain.Station;
 
 import javax.persistence.Column;
 import javax.persistence.Embedded;
