+++ src/main/java/nextstep/subway/line/dto/LineResponse.java
@@ -1,7 +1,7 @@
-package nextstep.subway.dto;
+package nextstep.subway.line.dto;
 
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.Station;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.station.domain.Station;
 
 import java.util.List;
