+++ src/main/java/nextstep/subway/section/domain/Sections.java
@@ -1,4 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.section.domain;
+
+import nextstep.subway.station.domain.Station;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
