+++ src/test/java/nextstep/subway/section/SectionAcceptanceTest.java
@@ -3,10 +3,10 @@ package nextstep.subway.section;
 import io.restassured.RestAssured;
 import io.restassured.response.ValidatableResponse;
 import nextstep.subway.AbstractAcceptanceTest;
-import nextstep.subway.dto.LineRequest;
-import nextstep.subway.dto.LineResponse;
-import nextstep.subway.dto.SectionRequest;
-import nextstep.subway.dto.StationResponse;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.section.dto.SectionRequest;
+import nextstep.subway.station.dto.StationResponse;
 import nextstep.subway.station.StationAcceptanceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
